+++ src/components/Novelty/Novelty.scss
@@ -1,0 +1,17 @@
+.Novelty{
+    display: flex;
+    margin: 2.4rem  4.8rem;
+    align-items: center;
+    &__description {
+        padding: 20px;
+        h2 {
+            font-size: 3.5rem;
+        }
+        p {
+            font-size: 2rem;
+        }
+    }
+    &__img {
+        border-radius: 4px;
+    }
+}
